Do not orphan user connector data
Since users can destroy themselves through the edit page, and schools
can be removed from the system by the admins - it is necessary to verify
that the related user_connector records are removed if either of these
actions occur. This is to prevent orphaned data from staying in the
database.